fix tx
