+++ internal/repository/content_value.go
@@ -9,6 +9,7 @@ import (
 type ContentValueRepo interface {
 	base.CRUDRepository[model.ContentValue]
 	FindByContentID(cID uint) ([]model.ContentValue, error)
+	WithTx(tx *gorm.DB) ContentValueRepo
 }
 
 type contentValueRepository struct {
@@ -21,6 +22,10 @@ func NewContentValueRepository(db *gorm.DB) ContentValueRepo {
 		BaseRepository: base.NewBaseRepository[model.ContentValue](db),
 		db:             db,
 	}
+}
+
+func (r *contentValueRepository) WithTx(tx *gorm.DB) ContentValueRepo {
+	return NewContentValueRepository(tx)
 }
 
 func (r *contentValueRepository) FindByContentID(cID uint) ([]model.ContentValue, error) {
